Potential fix for code scanning alert no. 14: Multiplication result converted to larger type

+++ arch/x86/events/intel/uncore_discovery.c
@@ -298,7 +298,7 @@ static int __parse_discovery_table(resource_size_t addr, int die,
 	}
 	iounmap(io_addr);
 
-	size = (1 + global.max_units) * global.stride * 8;
+	size = (unsigned long)(1 + global.max_units) * global.stride * 8;
 	io_addr = ioremap(addr, size);
 	if (!io_addr)
 		return -ENOMEM;
